Use `u32` as line width
When using `u128` internally in `wrap_optimal_fit`, we can let the
line width go all the way up to `u32` without risking overflows.

+++ benches/linear.rs
@@ -9,7 +9,7 @@ use criterion::{criterion_group, criterion_main};
 
 use lipsum::lipsum_words_from_seed;
 
-const LINE_LENGTH: u16 = 60;
+const LINE_LENGTH: u32 = 60;
 
 /// Generate a lorem ipsum text with the given number of characters.
 fn lorem_ipsum(length: usize) -> String {
