+++ fuzz/fuzz_targets/fill_optimal_fit.rs
@@ -3,7 +3,7 @@ use libfuzzer_sys::fuzz_target;
 use textwrap::wrap_algorithms;
 use textwrap::Options;
 
-fuzz_target!(|input: (String, u16)| {
+fuzz_target!(|input: (String, u32)| {
     let options = Options::new(input.1).wrap_algorithm(wrap_algorithms::OptimalFit::default());
     let _ = textwrap::fill(&input.0, &options);
 });
